css-transforms-1/2d-rotate-js.html

<!DOCTYPE html>
<html>
	<head>
		<title>JS test: Rotate via javascript must show the correct computed rotation</title>
		<link rel="author" title="Rick Hurst" href="http://mrkn.co/axegs">
		<link rel="help" href="http://www.w3.org/TR/css-transforms-1/#transform-property">
		<meta name="flags" content="css3, rotate, svg">
		<meta name="assert" content="Asserting that you can rotate an element with JS and it show up in CSS computed values not as a matrix but as the rotation">
		<style>
			#box{
				margin-top:30px;
				display: block;
				width: 50px;
				height: 50px;
				background-color: green;
			}
		</style>
	</head>
    <body>
		<h1>Rotate via JS</h1>
        <div id="log"></div>
        <div id="box"></div>
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<script>
			(function(){
				var box = document.getElementById("box"),
					assertion = "",
					expectation = "matrix(0.866025403784439, 0.5, -0.5, 0.866025403784439, 0, 0)",
					should_string = "this should make a small green square rotated 30 degrees, and the browser should return the rotation as 30 degrees as the computed value for a matrix",
					test_function = {};
				box.style.transform = "rotate(30deg)";
				assertion = window.getComputedStyle(box).getPropertyValue("transform");
				test_function = function(){
					assert_equals(assertion, expectation);
				}		
				test(test_function, should_string);		
			})();	
		</script>
    </body>
</html>
